abstract twitch route functionality to twitch service
